feat: Extract attributeValue into helper function Used in all transforms
