+++ lib/ast-input-transform.js
@@ -2,6 +2,7 @@
 
 const checkAttributes = require('./helpers/check-attributes');
 const supportsAttribute = require('./helpers/supports-attribute');
+const expressionForAttributeValue = require('./helpers/expression-for-attribute-value');
 const getTag = require('./helpers/get-tag');
 
 const reLines = /(.*?(?:\r\n?|\n|$))/gm;
@@ -97,15 +98,6 @@ class AngleBracketInputPolyfill {
 
     // in order to debug in https://https://astexplorer.net/#/gist/0590eb883edfcd163b183514df4cc717
     // **** copy from here ****
-    function transformAttributeValue(attributeValue) {
-      switch (attributeValue.type) {
-        case 'TextNode':
-          return b.string(attributeValue.chars);
-        case 'MustacheStatement':
-          return b.path(attributeValue.path);
-      }
-    }
-
     let visitor = {
       ElementNode(node) {
         let tag = getTag(node, sourceLines);
@@ -119,8 +111,8 @@ class AngleBracketInputPolyfill {
             tag === 'Textarea'
               ? textareaAttributes
               : isCheckbox
-              ? checkboxAttributes
-              : inputAttributes;
+                ? checkboxAttributes
+                : inputAttributes;
 
           checkAttributes(node, supportedAttributes, moduleName);
 
@@ -129,15 +121,16 @@ class AngleBracketInputPolyfill {
             .map(attribute => Object.assign({}, attribute, { name: attribute.name.slice(1) }));
           let attrs = attributes
             .filter(({ name }) => supportsAttribute(name, supportedAttributes))
-            .map(attribute =>
-              attributeToPropertyMap[attribute.name]
-                ? Object.assign({}, attribute, { name: attributeToPropertyMap[attribute.name] })
-                : attribute
+            .map(
+              attribute =>
+                attributeToPropertyMap[attribute.name]
+                  ? Object.assign({}, attribute, { name: attributeToPropertyMap[attribute.name] })
+                  : attribute
             );
 
           let hash = b.hash(
             [...props, ...attrs].map(({ name, value, loc }) =>
-              b.pair(name, transformAttributeValue(value), loc)
+              b.pair(name, expressionForAttributeValue(b, value), loc)
             )
           );
