+++ lib/ast-link-to-transform.js
@@ -2,6 +2,7 @@
 
 const checkAttributes = require('./helpers/check-attributes');
 const supportsAttribute = require('./helpers/supports-attribute');
+const expressionForAttributeValue = require('./helpers/expression-for-attribute-value');
 const SilentError = require('silent-error');
 
 const reservedProps = ['@route', '@model', '@models', '@query'];
@@ -31,21 +32,6 @@ class AngleBracketLinkToPolyfill {
 
     // in order to debug in https://astexplorer.net/#/gist/0590eb883edfcd163b183514df4cc717/6284c6222fe671a1be16432477cdf83c840dfcf0
     // **** copy from here ****
-    function transformAttributeValue(attributeValue) {
-      switch (attributeValue.type) {
-        case 'TextNode':
-          return b.string(attributeValue.chars);
-        case 'MustacheStatement':
-          if (attributeValue.params.length) {
-            // Deals with helper usage
-            return b.path(
-              b.sexpr(attributeValue.path.original, attributeValue.params, null, attributeValue.loc)
-            );
-          }
-          return b.path(attributeValue.path);
-      }
-    }
-
     let visitor = {
       ElementNode(node) {
         if (node.tag.toLowerCase() === 'linkto') {
@@ -72,17 +58,17 @@ class AngleBracketLinkToPolyfill {
 
           if (route) {
             if (needsParamsHelper) {
-              helperParams.push(b.pair('route', transformAttributeValue(route.value)));
+              helperParams.push(b.pair('route', expressionForAttributeValue(b, route.value)));
             } else {
-              params.push(transformAttributeValue(route.value));
+              params.push(expressionForAttributeValue(b, route.value));
             }
           }
 
           if (model) {
             if (needsParamsHelper) {
-              helperParams.push(b.pair('model', transformAttributeValue(model.value)));
+              helperParams.push(b.pair('model', expressionForAttributeValue(b, model.value)));
             } else {
-              params.push(transformAttributeValue(model.value));
+              params.push(expressionForAttributeValue(b, model.value));
             }
           }
 
@@ -90,7 +76,7 @@ class AngleBracketLinkToPolyfill {
             if (models.value.path.original === 'array') {
               params.push(...models.value.params);
             } else {
-              helperParams.push(b.pair('models', transformAttributeValue(models.value)));
+              helperParams.push(b.pair('models', expressionForAttributeValue(b, models.value)));
             }
           }
 
@@ -98,7 +84,7 @@ class AngleBracketLinkToPolyfill {
             if (query.value.path.original === 'hash') {
               params.push(b.sexpr('query-params', null, query.value.hash, query.loc));
             } else {
-              helperParams.push(b.pair('query', transformAttributeValue(query.value)));
+              helperParams.push(b.pair('query', expressionForAttributeValue(b, query.value)));
             }
           }
 
@@ -116,7 +102,7 @@ class AngleBracketLinkToPolyfill {
 
           let hash = b.hash(
             [...props, ...attrs].map(({ name, value, loc }) =>
-              b.pair(name, transformAttributeValue(value), loc)
+              b.pair(name, expressionForAttributeValue(b, value), loc)
             )
           );
