+++ lib/ast-transform.js
@@ -2,6 +2,7 @@
 
 const getTag = require('./helpers/get-tag');
 const sourceForNode = require('./helpers/source-for-node');
+const expressionForAttributeValue = require('./helpers/expression-for-attribute-value');
 
 const reLines = /(.*?(?:\r\n?|\n|$))/gm;
 const ALPHA = /[A-Za-z]/;
@@ -31,25 +32,6 @@ class AngleBracketPolyfill {
 
     function replaceNestedComponents(string) {
       return string.replace(/::/g, '/');
-    }
-
-    function isSimple(mustache) {
-      return mustache.params.length === 0 && mustache.hash.pairs.length === 0;
-    }
-
-    function expressionForAttributeValue(value) {
-      if (value.type === 'TextNode') {
-        return b.string(value.chars);
-      } else if (value.type === 'MustacheStatement') {
-        // TODO: Resolve ambiguous case data-foo="{{is-this-a-helper}}"
-        if (isSimple(value)) {
-          return value.path;
-        } else {
-          return b.sexpr(value.path, value.params, value.hash, value.loc);
-        }
-      } else if (value.type === 'ConcatStatement') {
-        return b.sexpr('concat', value.parts.map(expressionForAttributeValue));
-      }
     }
 
     function getSelfClosing(element) {
@@ -125,7 +107,9 @@ class AngleBracketPolyfill {
           'hash',
           [],
           b.hash(
-            attrs.map(attr => b.pair(attr.name, expressionForAttributeValue(attr.value), attr.loc))
+            attrs.map(attr =>
+              b.pair(attr.name, expressionForAttributeValue(b, attr.value), attr.loc)
+            )
           )
         );
       }
@@ -196,7 +180,7 @@ class AngleBracketPolyfill {
 
         let hash = b.hash(
           args.map(arg =>
-            b.pair(arg.name.slice(1), expressionForAttributeValue(arg.value), arg.loc)
+            b.pair(arg.name.slice(1), expressionForAttributeValue(b, arg.value), arg.loc)
           )
         );
